feat: implement reusable OrganizationCard component with structured sections
