+++ src/components/common/organization-card/sections/index.ts
@@ -1,0 +1,7 @@
+export { ContactsSection } from "./ContactsSection";
+export { DepartmentsSection, InlineListSection } from "./ListSections";
+export { EventsSection } from "./EventsSection";
+export { GallerySection } from "./GallerySection";
+export { JoinCta } from "./JoinCta";
+export { OrganizationCardHeader } from "./Header";
+export { PresentationSection, RichTextSection } from "./TextSections";
